Move logic from Job controller to new Job  classes
See Redmine ticket #11487.
